cros: Window cycle list (alt-tab) resets if open during desk swap.
While the flag for limiting windows to the current active desk is
enabled, if a user swaps desks while the window cycle list is open, the
contents will not update to the new active desk.

This CL makes the window cycle list reset if it is open and a desk
swap occurs while the aforementioned flag is enabled.

Test: LimitedWindowCycleControllerTest.CycleShowsActiveDeskWindows
Bug: 1067327
Change-Id: I9d940ad82a2f05ab1d2fd9e8c5115d5f3eda9826
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2398559
Commit-Queue: Jeremy Chinsen <chinsenj@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806749}