Add IPH Snooze service
Add FeaturePromoSnoozeService that handles the dismiss and snooze of IPH
bubble by the user. This service uses PrefService as the storage for
snoozed IPH status.

This service is a middleware between IPH frontend and backend. The
frontend asks it for IPH display permission before asking the backend.
The frontend should notify this service when the user clicks dismiss or
snooze button on the IPH bubble.

Note that in addition to snooze, this service handles dismiss.
Otherwise, we will have to change backend to incorporate dismiss logic.

Bug: 1121399
Change-Id: Ie0817a1ec15ff835344aa95dc17a27c0ae6f4338
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2398979
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Collin Baker <collinbaker@chromium.org>
Reviewed-by: Wei Li <weili@chromium.org>
Commit-Queue: Keren Zhu <kerenzhu@chromium.org>
Auto-Submit: Keren Zhu <kerenzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806765}