[Nearby] Prevent deadlock when using BluetoothServerSocket.
BluetoothServerSocket is created on the main thread, but its public
methods are used on a separate dedicated thread. mojo::Remote
objects (namely, BluetoothServerSocket's |server_socket_|) must be
bound on the same thread they are used on, to prevent deadlock. So,
we hold onto this mojo::PendingRemote |pending_server_socket_| until
Accept() is called, at which point |server_socket_| is bound with
it (it is acceptable to pass a mojo::PendingRemote around multiple
threads).

bluetooth_server_socket_unittest.cc has been adjusted for this change.
I have verified that the test fails if this fix is not present.

Further, mojo::Remotes expect to be run on a SequencedTaskRunner.
Tweak our ImplementationPlatform::CreateMultiThreadExecutor() impl
to provide a SequencedTaskRunner to satisfy this requirement.

Bug: 1126620
Change-Id: I5dbb11d0d97b8167f60acb08fc1ae962bdf17f06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2407296
Reviewed-by: James Vecore <vecore@google.com>
Commit-Queue: Ryan Hansberry <hansberry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806786}