background sync: Deflake stop-with-registered-one-shot-sync-task.html
This test starts a SW and uses an internals API to stop it.

When a SW stops while it has an inflight request, it calls the request
callback with kServiceWorkerFailed status in
ServiceWorkerVersion::OnStoppedInternal():

  base::IDMap<std::unique_ptr<InflightRequest>>::iterator iter(
      &inflight_requests_);
  while (!iter.IsAtEnd()) {
    std::move(iter.GetCurrentValue()->error_callback)
        .Run(blink::ServiceWorkerStatusCode::kErrorFailed);
...

BackgroundSyncManager::GetEventStatusString() was not expecting this
status code, so crashed.

Bug: 1098832
Change-Id: I7d0da49a5b369f23062b27cebbc656097bda2ed0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2405045
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806822}