Intersect presentation area with the scrollable area
When determining the presentation area for delegated ink trails, its
possible that the canvas that the user has provided is larger than the
viewport of the browser window. In this case, the presentation area
should intersect with the scrollable area so that it can extend to the
scroll bars, but not beyond them. This change also ensures that the
presentation area is clipped to the edge of iframes as well.

Bug: 1052145
Change-Id: I349616b25a0523e825dffb37c19e14fe737dc3f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2341949
Reviewed-by: Robert Flack <flackr@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806824}