Add deprecation message for allow attribute.
The use of commas as separators in the iframe allow attribute is not
allowed, according to the specification, but it was previously parsed,
as it can appear in HTTP headers. This CL adds a deprecation message,
and is the first step towards removing this.

Blink intent thread:
https://groups.google.com/a/chromium.org/d/msg/blink-dev/43EhaSQWm6M/iqvQ3hY0AwAJ

Bug: 1062400
Change-Id: I2efff7c9610ca7b85145a7600d1c0074ad2e30be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2313298
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Commit-Queue: Ian Clelland <iclelland@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806885}