Update magnifier focus for focus changes in Arc windows.
This requires intents to ensure the ArcAccessibilityHelperService
is enabled in Android, and the editability of the control to be passed
through to the MagnificationManager.

To work with no other accessibility features are enabled, this also
requires the Android side change: http://ag/12532278

This change will still be safe to merge without that change, as the
intents will just be ignored.

BUG=b:167133958
TEST=Manual (tab around apps and observe magnifier,
enable and disable magnifier and other accessibility features)
R=hirokisato@chromium.org, sarakato@chromium.org
AX-Relnotes: Magnifier now follows focus in Android apps.

Change-Id: I2fd1a7145be3475a9c288715c4d442f6bc035c57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2389802
Commit-Queue: Matthew Duggan <mduggan@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: David Jacobo <djacobo@chromium.org>
Reviewed-by: Sara Kato <sarakato@chromium.org>
Reviewed-by: Hiroki Sato <hirokisato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806904}