[Sharesheet] Add temp icon from drive share action.
Currently the drive share action returns an empty icon, however this
will no pass a size DCHECK. The reason we return an empty icon is
because there is some refactoring needed for the GetActionIcon method
and we should not create the icon at this stage (crbug.com/1127750).
However because we don't want to cause a DCHECK failure, and also there
is already another implementation that does the same thing, so I will
add this temp solution and this will be fixted together.

BUG=1097623

Change-Id: Iff4fcaac6aa0d2308d9b9539f861303164346f9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2409850
Reviewed-by: Melissa Zhang <melzhang@chromium.org>
Commit-Queue: Maggie Cai <mxcai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806917}