[Mfill][Android] Recreate destroyed filling components
There are rare cases when the the Android native UI is recreated but the
native parts remain untouched (e.g. when the theme is changed).

To prevent native classes using stale providers, reset them when the
component is destroyed and recreate them lazily as needed.
(Since this touches the way the native object is created, clean that
up as well to prevent rare crashes.)

Bug: 1094397, 1051030
Change-Id: I9c5cd1e225db1b4ca61c5136571a0e552fe86940
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2403641
Reviewed-by: Vadym Doroshenko  <dvadym@chromium.org>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806934}