arc: bluetooth: Avoid useless GATT disconnected log
Currently OnGattDisconnected() is called every time DeviceRemoved() is
triggered, and will output a warning log if we doesn't hold a GATT
connection for that device, but actually this should be a normal case.
This patch adds an early return to avoid the log in that case.

BUG=None
TEST=Running system for a while, no more "OnGattDisconnected called, but
  no gatt connection was found" in chrome log.

Change-Id: Ia3dcc4e65cbc1e66ef1f7c34d51d2940a28095ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2409660
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Auto-Submit: Jie Jiang <jiejiang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806935}