Delete PermissionRequest::GetEmbeddingOrigin.
PermissionRequest:GetEmbeddingOrigin was not implemented for all
permission types, which made it difficult to use anywhere except when
constructing the prompt strings. This patch moves that construction into
UI code where the strings are used, which also simplifies the plumbing
in PermissionPromptImpl. This patch also moves the Flash warning to use
the new codepath, which allows us to delete
PermissionRequest::GetMessageTextWarningFragment.

This has a minor visual impact on the permission prompt bubble on
Desktop. See the associated bug for screenshots.

Bug: 1110905
Change-Id: Icfeedd011d620e467a5dbbf75bea6875f4af9379
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2379729
Reviewed-by: Andy Paicu <andypaicu@chromium.org>
Reviewed-by: Olesia Marukhno <olesiamarukhno@google.com>
Commit-Queue: Bret Sepulveda <bsep@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806960}