Enable AutoColorReadability for notification text and icons
This enables automatic color contrast checks in these notification views
so that we can use any theme color in a follow-up CL:
- NotificationControlButtonsView
- NotificationHeaderView
- NotificationViewMD

This will make sure that the text and icons are at least readable even
if the colors don't combine well with each other. Note that this will
still render the main notification content text too bright as it already
expect the background to be set to the native theme color.

Bug: 1113313
Change-Id: Ic1bc0816eeb6242b93a1195c2c03aff81726f83c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2398507
Commit-Queue: Richard Knoll <knollr@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806974}