Do not kill renderer for unexpected messages
We are not going to use this feature in the near future as there are
some bugs that make renderers send messages while cached. One of the
issues we are seeing is that in processes with active and cached pages
we can not guarantee that no messages are sent for the cached ones as
some of the task queues are shared between pages. Until we have
dedicated queues (the Multiple Blink Isolates project should help here)
we should not consider a message unexpected if it comes from a process
with non cached pages.

We only have one compositor task queue and currently it is
only frozen if all pages in the renderer are frozen, so if there is an
active page it can still process tasks for the frozen page. This is what
is causing the flakes in BackForwardCacheBrowserTest.WindowOpen

Also input task queues are not frozen when the page freezes so that is
causing flakes in
BackForwardCacheBrowserTest.NavigateBackForwardRepeatedly so this patch
disables the check for unexpected message in that test until the issue
is fixed.

Bug: 1116023
Bug: 1099395
Change-Id: Ibe4c89f14f828d581c1156fa2d1e8e63eeac1e39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2397698
Commit-Queue: Carlos Caballero <carlscab@google.com>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806996}