BackgroundFetchTestDataManager usability issue
Currently the data manager has a bind, making the test component unusable
when we have separate UI, IO threads, and the service worker is running
on the UI thread (under flag `kServiceWorkerOnUI`).
As the bind must happen on the IO thread, but
`BackgroundFetchContext::InitializeOnCoreThread` which calls this, must
happen on the UI thread.

This CL opens the issue up for discussion, to highlight the problem.
But the fix is not intended to be a solution.

Change-Id: I8565247dcd524336b725105cc83d2ddcb9ece1aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2404793
Commit-Queue: Clovis PJ <clovispj@google.com>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#807002}