Fix permission bubble icons not updating with theme changes.
This patch changes the request row icons and Flash's learn more button
to color-tracking versions of the basic Views, which fixes the bug. This
also changes the learn more button to track kColorId_DefaultIconColor
instead of kColorId_LabelEnabledColor, but this is more consistent and
matches the request row icon.

Bug: 1125510
Change-Id: I0931622efbf3a4a6e56ad236a281f4201b9d27d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2409958
Commit-Queue: Bret Sepulveda <bsep@chromium.org>
Reviewed-by: Olesia Marukhno <olesiamarukhno@google.com>
Cr-Commit-Position: refs/heads/master@{#807034}