Make WebUIURLLoaderFactory always owned by its |receivers_|.
This CL changes (and slightly renames) CreateWebUIURLLoaderFactory so
that it returns mojo::PendingRemote to a WebUIURLLoaderFactory that is
owned by its |receivers_| and will self-delete when the last receiver
disconnects.

This CL removes the ability to directly construct and own a
std::unique_ptr<WebUIURLLoaderFactory>, because this ability means that
the factory can be destructed while receivers bound via the Clone method
are still alive (see the associated bug).

This CL mostly just follows the pattern established earlier by
https://crrev.com/c/2337411.

Bug: 1106995
Change-Id: I6079a8b4b3a17d838539a62234b788a3353c76c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2358309
Reviewed-by: Sean Topping <seantopping@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Reviewed-by: James MacLean <wjmaclean@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/master@{#807050}