weblayer: Set compositor background wrt dark mode
Set the background color on ContentViewRenderView according to the dark
mode, ie set it to black if dark mode is enabled.

Then also set the same background on the compositor which is the the
part that actually appears to make a difference visually.

This is meant to be merged back to 86.

Bug: 1127873
Change-Id: Ifcb9fa47af1e578b963f3f47ba5dc2b8252b80b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2412189
Commit-Queue: Bo <boliu@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#807122}