Download home: Fix a flaky test with RecyclerView issue.
In devices with small screen, "page 1" can fall outside of the view
port. We have to scroll to make the recycler view create the view
holder in memory.

About view port:
1. RecyclerView won't create view holder outside view port to save memory.
2. IsDisplay() won't check things outside view port.
3. In order to verify stuff outside view port, we have to scroll to a
particular view holder or position.

To scroll, there are a couple of ways:
1. Use any of 3 RecyclerViewActions.* API.
2. Implement our own ViewAction for RecyclerView, and scroll with the recycler view handle.

Bug: 1039491
Change-Id: I1669370abf595140313fc8cfcf2728e12e3b3389
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2411557
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Commit-Queue: Xing Liu <xingliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#807128}