Pass in ClientPhishingRequest as a unique_ptr.
Previously, the ClientPhishingRequest was passed
as a raw pointer into SendClientPhishingRequest
and then converted back to a unique_ptr when
passed into StartClientPhishingRequest. When
using a raw pointer,it has to be made sure
to be deleted otherwise there is a memory leak.
To make this easier to maintain, it is switched
to a unique_ptr.

Bug: 1122226
Change-Id: Ia344ea26661eec202cd27224889ba398f8264835
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2379252
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Commit-Queue: Bettina Dea <bdea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#807150}