MacOS zero copy capture: Add OnIncomingCapturedExternalBuffer
On macOS, a GpuMemoryBuffer pool (CVPixelBufferPool, to be specific)
is managed internally by AVCaptureVideoDataOutput. It will burp up
a new CMSampleBuffer (which may have a CVPixelBuffer, which may have
an IOSurface) every frame. In the case that there is an IOSurface,
we will want to send that surface to the receiver (encoder, compositor,
etc), rather than copy it a bunch of times.

This does not mesh well with the VideoCaptureDevice structure,
where the pool of buffers is managed by the VideoCapturePool, and
handed to the capturer to populate.

Fix this situation with the following changes.

To VideoCapturePool, add a method to reserve a buffer id that will not
conflict with other buffer ids from the pool. External buffers will
use these ids.

To VideoCaptureDevice::Client add a OnIncomingCapturedExternalBuffer
method, which takes a gfx::GpuMemoryBufferHandle. This method will
scope the external buffer's lifetime in the receiver to the receiver's
use of the buffer.

Bug: 1125879
Change-Id: Ic6cfb7e0e91a97d4489b37cfc7e013b91db369f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2406717
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Markus Handell <handellm@google.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#807159}